:bug: Projet - Correctif des regressions suite aux merges
