+++ Site/index.php
@@ -11,15 +11,16 @@ require_once PATH_MODELS.'paysDAO.php';
 session_start();
 
 if(empty($_GET["lang"])) {
-    if (isset($_SESSION["lang"]))
+    if (isset($_SESSION["lang"]) && !empty($_SESSION["lang"]))
         $lang = $_SESSION["lang"];
     else
         $lang = getLangage();
 }
-else
+else {
     $lang = setLangageById($_GET["lang"]);
+}
 
-require_once(PATH_TEXTES.$lang.'.php');
+require_once PATH_TEXTES.$lang.'.php';
 
 //vérification de la page demandée
 if(isset($_GET['page']))
